+++ pepclibs/msr/HWPRequestPkg.py
@@ -73,7 +73,7 @@ class HWPRequestPkg(_FeaturedMSR.FeaturedMSR):
 
     regaddr = MSR_HWP_REQUEST_PKG
     regname = "MSR_HWP_REQUEST_PKG"
-    vendor = "GenuineIntel"
+    vendor_name = "GenuineIntel"
 
     def __init__(self,
                  cpuinfo: CPUInfo.CPUInfo,