+++ pepclibs/msr/TurboRatioLimit.py
@@ -106,7 +106,7 @@ class TurboRatioLimit(_FeaturedMSR.FeaturedMSR):
 
     regaddr = MSR_TURBO_RATIO_LIMIT
     regname = "MSR_TURBO_RATIO_LIMIT"
-    vendor = "GenuineIntel"
+    vendor_name = "GenuineIntel"
 
     def __init__(self,
                  cpuinfo: CPUInfo.CPUInfo,