+++ pepclibs/msr/TurboRatioLimit1.py
@@ -56,7 +56,7 @@ class TurboRatioLimit1(_FeaturedMSR.FeaturedMSR):
 
     regaddr = MSR_TURBO_RATIO_LIMIT1
     regname = "MSR_TURBO_RATIO_LIMIT1"
-    vendor = "GenuineIntel"
+    vendor_name = "GenuineIntel"
 
     def __init__(self,
                  cpuinfo: CPUInfo.CPUInfo,