+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ## [ADD NEW VERSION HERE] - ADD DATE HERE
 ### Fixed
  - pepc tpmi * --base: Do not fail when 'tpmi_info' feature (0x81) is missing.
+ - pepc: Fix regression in 1.6.19: MSR modules did not work.
 ### Added
 ### Removed
 ### Changed